+++ csvkit/utilities/csvsql.py
@@ -69,8 +69,9 @@ class CSVSQL(CSVKitUtility):
             '--db-schema', dest='db_schema',
             help='Optional name of database schema to create table(s) in.')
         self.argparser.add_argument(
-            '-y', '--snifflimit', dest='sniff_limit', type=int,
-            help='Limit CSV dialect sniffing to the specified number of bytes. Specify "0" to disable sniffing.')
+            '-y', '--snifflimit', dest='sniff_limit', type=int, default=1024,
+            help='Limit CSV dialect sniffing to the specified number of bytes. '
+                 'Specify "0" to disable sniffing entirely, or "-1" to sniff the entire file.')
         self.argparser.add_argument(
             '-I', '--no-inference', dest='no_inference', action='store_true',
             help='Disable type inference when parsing the input.')
@@ -168,12 +169,13 @@ class CSVSQL(CSVKitUtility):
                     table_name = os.path.splitext(os.path.basename(f.name))[0]
 
             table = None
+            sniff_limit = self.args.sniff_limit if self.args.sniff_limit != -1 else None
 
             try:
                 table = agate.Table.from_csv(
                     f,
                     skip_lines=self.args.skip_lines,
-                    sniff_limit=self.args.sniff_limit,
+                    sniff_limit=sniff_limit,
                     column_types=self.get_column_types(),
                     **self.reader_kwargs
                 )